[ADD] Created validity checker for Book objects

+++ src/main/java/com/revature/p1/app/services/BookService.java
@@ -19,19 +19,32 @@ public class BookService {
         return bookDAO.findAll();
     }
 
-    public List<Book> getBookById(String book) {
-        return bookDAO.getById(book);
+    public List<Book> getBookById(String uuid) {
+        if(uuid.length() != 36) return null;
+        return bookDAO.getById(uuid);
     }
 
     public boolean insertBook(Book book) {
+        if(!isBookValid(book)) return false;
         return bookDAO.insert(book);
     }
 
     public boolean updateBook(Book book) {
+        if(!isBookValid(book)) return false;
         return bookDAO.update(book);
     }
 
     public boolean deleteBook(Book book) {
+        if(!isBookValid(book)) return false;
         return bookDAO.delete(book);
+    }
+
+    public boolean isBookValid(Book book) {
+        if (book == null) return false;
+        if (book.getTitle() == null || book.getTitle().trim().equals("")) return false;
+        if (book.getAuthor() == null || book.getAuthor().trim().equals("")) return false;
+        if (book.getPageCount() == null || book.getPageCount() < 0) return false;
+
+        return true;
     }
 }
